models/2.chatmodels/chatmodel_anthropic.py
from langchain_anthropic import ChatAnthropic
from dotenv import load_dotenv

load_dotenv()

model = ChatAnthropic(model='claude-3-5-sonnet-20241022')

result = model.invoke('What is the capital of India')

print(result.content)
"""First generate api key for Anthropic and set it in .env file as ANTHROPIC_API_KEY.
"""